// projects/viescloud-utils/src/lib/model/Popup.model.ts
export const POPUP_DATA = 'POPUP_DATA';
export const POPUP_DISMISS = 'POPUP_DISMISS';

export enum PopupType {
    NONE = "NONE",
    STRING_SNACKBAR = "STRING_SNACKBAR",
    DYNAMIC_STRING_SNACKBAR = "DYNAMIC_STRING_SNACKBAR",
    LOADING_DIALOG = "LOADING_DIALOG",
    MESSAGE_POPUP = "MESSAGE_POPUP",
    DYNAMIC_MESSAGE_POPUP = "DYNAMIC_MESSAGE_POPUP"
}